requirements.txt
asgiref==3.3.1
<<<<<<< HEAD
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
=======
certifi==2020.11.8
chardet==3.0.4
>>>>>>> 012d00e9e5789795c2ae4ff2ca3197f5457bb895
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.3
<<<<<<< HEAD
django-dotenv==1.4.2
=======
>>>>>>> 012d00e9e5789795c2ae4ff2ca3197f5457bb895
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
<<<<<<< HEAD
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.7
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
=======
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.7
psycopg2==2.8.6
PyJWT==1.7.1
>>>>>>> 012d00e9e5789795c2ae4ff2ca3197f5457bb895
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
<<<<<<< HEAD
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
=======
sqlparse==0.4.1
>>>>>>> 012d00e9e5789795c2ae4ff2ca3197f5457bb895
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
